add author

+++ resources/views/components/comment.blade.php
@@ -3,9 +3,11 @@
         @csrf
         <h4>{{ $comment->user->name }}</h4>
         <input type="hidden" value="{{ $comment->id }}" name="id" />
-        <textarea name="body" {{ auth()->user()->id !== $comment->user_id ? 'disabled' : 'required' }} rows="3"
+        <textarea name="body" {{ auth()->user() && auth()->user()->id !== $comment->user_id ? 'disabled' : 'required' }} rows="3" 
             cols="10" style="width: 100%">{{ $comment->body }}</textarea>
-        <a href="/deletecomment/{{ $comment->id }}" class="btn btn-danger">Delete</a>
-        <button type="submit" class="btn btn-primary">Update</button>
+        @if(auth()->user() && auth()->user()->id === $comment->user_id)
+            <a href="/deletecomment/{{ $comment->id }}" class="btn btn-danger">Delete</a>
+            <button type="submit" class="btn btn-primary">Update</button>
+        @endif
     </form>
 @endforeach
